integrate processIncomingMessages. Author: zrd1

+++ app/message-processors/lib/processors/include/messageProcessor.h
@@ -4,7 +4,7 @@ namespace MessageProcessors {
 
 // from the client
 struct RequestMessageDTO {
-    int clientId;
+    uintptr_t clientId;
     std::string command;
     std::string data;
 };
